pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "dengue-dashboard"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
  "fastapi",
  "uvicorn[standard]",
  "python-dotenv",
  "pandas",
  "supabase",
  "psycopg2-binary",
  "requests",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["api*", "worker*", "denguard*"]